Revert "Попытка вынести роуты в отдельный файл."
This reverts commit eba7948953ca9f7ad751ed59e03430bb4913a172.
