+++ app.js
@@ -1,16 +1,47 @@
+const express = require('express');
 const mongoose = require('mongoose');
 const helmet = require('helmet');
+const { errors } = require('celebrate');
+
 const limiter = require('./middlewares/rateLimiter');
-const app = require('./routes/index');
+
+const routeSignup = require('./routes/signup');
+const routeSignin = require('./routes/signin');
+
+const auth = require('./middlewares/auth');
+
+const routeUsers = require('./routes/users');
+const routeCards = require('./routes/cards');
+
+const NotFoundError = require('./errors/NotFoundError');
+const errorHandler = require('./middlewares/errorHandler');
 
 const URL = 'mongodb://127.0.0.1:27017/mestodb';
 const { PORT = 3000 } = process.env;
 
 mongoose.set('strictQuery', true);
 
+mongoose.connect(URL);
+
+const app = express();
+
 app.use(helmet());
+
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+
 app.use(limiter);
 
-mongoose.connect(URL);
+app.use('/', routeSignup);
+app.use('/', routeSignin);
+
+app.use(auth);
+
+app.use('/users', routeUsers);
+app.use('/cards', routeCards);
+
+app.use((req, res, next) => next(new NotFoundError('Страницы по запрошенному URL не существует')));
+app.use(errors());
+app.use(errorHandler);
 
 app.listen(PORT);
